changed types of av's to int for some part of there life

+++ AlphaZeroPytorch/include/game/game.hpp
@@ -56,7 +56,7 @@ namespace AlphaZero {
 			std::size_t operator()(std::pair<std::shared_ptr<GameState>, std::vector<int>> const& s) const noexcept;
 		};
 		// optimization function its not a problem if not all are found
-		std::vector<std::pair<std::shared_ptr<AlphaZero::Game::GameState>, std::vector<float>>> identities(std::shared_ptr<GameState> state, std::vector<float>& actionProbs);
+		std::vector<std::pair<std::shared_ptr<AlphaZero::Game::GameState>, std::vector<int>>> identities(std::shared_ptr<GameState> state, std::vector<int>& actionProbs);
 
 		class Game {
 		public: std::tuple<int, int> BoardShape = { 3,3 };
@@ -74,7 +74,7 @@ namespace AlphaZero {
 			AlphaZero::Game::Game* game = new AlphaZero::Game::Game();
 
 			while (!game->state->done) {
-				std::vector<float> vec = { 0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 17.0, 18.0, 19.0, 20.0, 21.0, 22.0, 23.0, 24.0, 25.0, 26.0, 27.0, 28.0, 29.0, 30.0, 31.0, 32.0, 33.0, 34.0, 35.0, 36.0, 37.0, 38.0, 39.0, 40.0, 41.0 };
+				std::vector<int> vec = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41 };
 				auto idents = identities(game->state, vec);
 				idents[1].first->render();
 				std::cout << "your action: ";
